+++ sdk/formrecognizer/azure-ai-formrecognizer/azure/ai/formrecognizer/aio/_form_recognizer_client_async.py
@@ -90,10 +90,14 @@ class FormRecognizerClient(FormRecognizerClientBaseAsync):
         :keyword int polling_interval: Waiting time between two polls for LRO operations
             if no Retry-After header is present. Defaults to 5 seconds.
         :keyword str continuation_token: A continuation token to restart a poller from a saved state.
+        :keyword str locale: Locale of the receipt. Supported locales include: en-AU, en-CA, en-GB, en-IN,
+            en-US (default).
         :return: An instance of an AsyncLROPoller. Call `result()` on the poller
             object to return a list[:class:`~azure.ai.formrecognizer.RecognizedForm`].
         :rtype: ~azure.core.polling.AsyncLROPoller[list[~azure.ai.formrecognizer.RecognizedForm]]
         :raises ~azure.core.exceptions.HttpResponseError:
+        .. versionadded:: v2.1-preview
+            The *locale* keyword argument
 
         .. admonition:: Example:
 
@@ -145,10 +149,14 @@ class FormRecognizerClient(FormRecognizerClientBaseAsync):
         :keyword int polling_interval: Waiting time between two polls for LRO operations
             if no Retry-After header is present. Defaults to 5 seconds.
         :keyword str continuation_token: A continuation token to restart a poller from a saved state.
+        :keyword str locale: Locale of the receipt. Supported locales include: en-AU, en-CA, en-GB, en-IN,
+            en-US (default).
         :return: An instance of an AsyncLROPoller. Call `result()` on the poller
             object to return a list[:class:`~azure.ai.formrecognizer.RecognizedForm`].
         :rtype: ~azure.core.polling.AsyncLROPoller[list[~azure.ai.formrecognizer.RecognizedForm]]
         :raises ~azure.core.exceptions.HttpResponseError:
+        .. versionadded:: v2.1-preview
+            The *locale* keyword argument
 
         .. admonition:: Example: