+++ sdk/formrecognizer/azure-ai-formrecognizer/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 3.1.0b1 (Unreleased)
 
+**New features**
+
+- Recognize receipt methods now take keyword argument `locale` to optionally indicate the locale of the receipt for
+improved results
 
 ## 3.0.0 (2020-08-20)